+++ src/Violation/Renderer/Html/HtmlOutputRendererFactory.php
@@ -1,5 +1,5 @@
 <?php
-namespace SensioLabs\DeprecationDetector\Violation\Renderer\HtmlOutput;
+namespace SensioLabs\DeprecationDetector\Violation\Renderer\Html;
 
 
 use SensioLabs\DeprecationDetector\Violation\Renderer\MessageHelper\MessageHelper;
